nes: refactor: extract next-line predictor into its own class (#2210)
