+++ src/extension/xtab/node/xtabUtils.ts
@@ -3,7 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { Raw } from '@vscode/prompt-tsx';
 import { IResponsePart } from '../../../platform/chat/common/chatMLFetcher';
+import { toTextParts } from '../../../platform/chat/common/globalStringUtils';
 import { AsyncIterableObject } from '../../../util/vs/base/common/async';
 
 
@@ -59,4 +61,22 @@ export function linesWithBackticksRemoved(linesStream: AsyncIterableObject<strin
 
 		// ignore bufferedLine
 	});
+}
+
+export function constructMessages({ systemMsg, userMsg }: { systemMsg: string; userMsg: string }): Raw.ChatMessage[] {
+	return [
+		{
+			role: Raw.ChatRole.System,
+			content: toTextParts(systemMsg)
+		},
+		{
+			role: Raw.ChatRole.User,
+			content: toTextParts(userMsg)
+		}
+	] satisfies Raw.ChatMessage[];
+}
+
+export function charCount(messages: Raw.ChatMessage[]): number {
+	const promptCharCount = messages.reduce((total, msg) => total + msg.content.reduce((subtotal, part) => subtotal + (part.type === Raw.ChatCompletionContentPartKind.Text ? part.text.length : 0), 0), 0);
+	return promptCharCount;
 }
